Added a round counter process
state_machine_process stays sets and holds round_number at 0 when clr is set. An en bit is added to prevent moving to next state on the next rising edge if the clr bit is unset at the middle of a clock cycle. This way, state 0 gets a full clock cycle to run.

The state machine first counts up from 0 to 10. It wraps around 0 if enc
is set (for encryption) or moves to 10, counts down to 0 and cycles if
enc is unset.